requirements.txt
 streamlit
langchain-core
faiss-cpu
sentence-transformers
unstructured
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-groq
python-dotenv

